+++ lib/src/properties.dart
@@ -3,17 +3,39 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/rendering.dart';
 
 class GlobalToggleProperties<T> {
+  /// The position of the indicator relative to the indices of the values.
   final double position;
+
+  /// The current value which is given to the switch.
+  ///
+  /// Helpful if the value is generated e.g.
+  /// when the switch constructor is called.
   final T current;
+
+  /// The previous value of the switch.
   final T? previous;
+
+  /// The values which are given to the switch.
+  ///
+  /// Helpful if the list is generated e.g.
+  /// when the switch constructor is called.
+  final List<T> values;
+
+  /// The previous position of the indicator relative
+  /// to the indices of the values.
   final double previousPosition;
+
+  /// The [TextDirection] of the switch.
   final TextDirection textDirection;
+
+  /// The current [ToggleMode] of the switch.
   final ToggleMode mode;
 
   const GlobalToggleProperties({
     required this.position,
     required this.current,
     required this.previous,
+    required this.values,
     required this.previousPosition,
     required this.textDirection,
     required this.mode,
@@ -21,15 +43,28 @@ class GlobalToggleProperties<T> {
 }
 
 class DetailedGlobalToggleProperties<T> extends GlobalToggleProperties<T> {
+  /// The final width of the space between the icons.
+  ///
+  /// May differ from the value passed to the switch.
   final double dif;
+
+  /// The final size of the indicator.
+  ///
+  /// May differ from the value passed to the switch.
   final Size indicatorSize;
+
+  /// The size of the switch exclusive the outer wrapper
+  final Size switchSize;
 
   const DetailedGlobalToggleProperties({
     required this.dif,
     required this.indicatorSize,
+    required this.switchSize,
     required double position,
+    // TODO: Rename to current (small breaking change)
     required T value,
     required T? previousValue,
+    required List<T> values,
     required double previousPosition,
     required TextDirection textDirection,
     required ToggleMode mode,
@@ -37,6 +72,7 @@ class DetailedGlobalToggleProperties<T> extends GlobalToggleProperties<T> {
           position: position,
           current: value,
           previous: previousValue,
+          values: values,
           previousPosition: previousPosition,
           textDirection: textDirection,
           mode: mode,
@@ -44,7 +80,10 @@ class DetailedGlobalToggleProperties<T> extends GlobalToggleProperties<T> {
 }
 
 class LocalToggleProperties<T> {
+  /// The value.
   final T value;
+
+  /// The index of [value].
   final int index;
 
   const LocalToggleProperties({
@@ -54,6 +93,10 @@ class LocalToggleProperties<T> {
 }
 
 class AnimatedToggleProperties<T> extends LocalToggleProperties<T> {
+  /// A value between 0 and 1.
+  ///
+  /// 0 indicates that [value] is not selected.
+  /// 1 indicates that [value] is selected.
   final double animationValue;
 
   AnimatedToggleProperties.fromLocal({
@@ -79,7 +122,13 @@ class AnimatedToggleProperties<T> extends LocalToggleProperties<T> {
 }
 
 class RollingProperties<T> extends LocalToggleProperties<T> {
+  /// The size the icon should currently have.
   final Size iconSize;
+
+  /// Indicates if the icon is in the foreground.
+  ///
+  /// For [RollingIconBuilder] it indicates if the icon will be on the indicator
+  /// or in the background.
   final bool foreground;
 
   RollingProperties.fromLocal({
@@ -105,6 +154,7 @@ class RollingProperties<T> extends LocalToggleProperties<T> {
 }
 
 class SizeProperties<T> extends AnimatedToggleProperties<T> {
+  /// The size the icon should currently have.
   final Size iconSize;
 
   SizeProperties.fromAnimated({